+++ guide/LCO/config.js
@@ -1,11 +1,7 @@
+import { requireContextMap } from '_lib/utils'
 
-// https://medium.com/dailyjs/leveraging-webpack-power-to-import-all-files-from-one-folder-cddedd3201b3
-let files = require.context('./icons/', false, /\.svg$/);
-let reKey = /^(.+\/)*(.+)[.].+$/;
-let icons = files.keys().reduce((o, k) => ({ ...o, [k.replace(reKey, '$2')]: files(k).default }), {});
 
 // Exude Styleguide Library Config
-
 export default
 {
     background:
@@ -96,7 +92,7 @@ export default
     {
         named: 
         {
-            ...icons,
+            ...requireContextMap(require.context('./icons/', false, /\.svg$/), 'default'),
 
             caretDown: 'chevron_down',
             check: 'checkbox',